refactor: add loading folder
